Using threadpool2 as real thread pool
